Revert "dhcp: use landlock to remove ambient file access"
This reverts commit fbb7618c25b42a55de3a481bf027b50b7f15e43e.

landlock depends on psx. When cgo is enabled, psx uses a cgo
implementation, which is undesired in setups which use cgo
(e.g. scan2drive with turbojpeg), but don’t come with a libc:
the dhcp client doesn’t start, which means the device is no
longer reachable over the network.