Change how status is displayed
Currently the status is appended to the service status link and no label
is displayed if the service is running. The status stopped is displayed
with a warning color (yellow). To me this feels uneven and a bit chaotic
which makes it harder to get a good overview and find a service.

This does three things to improve the overview:
- Add a separate column for status label. This takes a bit of width away
  from the last log line.
- Color the stopped label less aggressive (gray/default instead of
  yellow/warning).
- Add a green label for running service.

Potential improvements for the future could be to show if a service is
enabled/supervised or adding actions like start/stop/restart/enable.