Always listen on gokrazy.HTTPUnixSocket (#296)
The change is made to allow other tools like breakglass or mkfs to
interact with the Gokrazy API when TLS is enabled. The tools currently
make HTTP requests to get system information or trigger reboots. When
TLS is enabled, these requests fail because Gokrazy redirects all HTTP
requests to HTTPS. While it is possible to configure the respective
clients to properly validate (self-signed) certificates, doing so is
cumbersome. Therefore, always listen on the Unix socket to avoid the
complexity of certificate validation. Username and password must still
be provided and will be validated.