do not read any password file when /etc/gokr-pw.txt is missing
See the comment for rationale. This makes it harder to take over a
gokrazy instance when having write access to /perm.

Whether the password should be read from /perm at all is another
valid question to ask. I am not sure anyone has such a setup?
If you are reading this and you do, please e-mail me.